backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "cohere>=5.20.0",
    "fastapi>=0.124.4",
    "openai-agents>=0.6.3",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "qdrant-client>=1.16.2",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "sentence-transformers>=5.2.0",
    "sqlalchemy>=2.0.45",
    "trafilatura>=2.0.0",
    "uvicorn>=0.38.0",
]
